refactor(model): split entities
